Move getCarColor to utils Fixes #40
